Move subrouters to routers directory
