+++ examples/next-hello-world/pages/api/trpc/[trpc].ts
@@ -2,7 +2,7 @@ import * as trpc from '@trpc/server';
 import { inferAsyncReturnType } from '@trpc/server';
 import * as trpcNext from '@trpc/server/adapters/next';
 import { z } from 'zod';
-import { postsRouter } from './posts';
+import { postsRouter } from '../../../routers/posts';
 
 // The app's context - is generated for each incoming request
 export async function createContext({
